Actu query3

+++ src/main/java/pe/edu/upc/waba/dtos/ComentarioDTO.java
@@ -9,7 +9,7 @@ import pe.edu.upc.waba.entities.Profesor;
 
 public class ComentarioDTO {
     private int idCo;
-    private float puntosCo;
+    private double puntosCo;
     private String resenaCo;
     private Profesor Pro;
 
@@ -21,7 +21,7 @@ public class ComentarioDTO {
         this.idCo = idCo;
     }
 
-    public float getPuntosCo() {
+    public double getPuntosCo() {
         return puntosCo;
     }
